backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.9
Pillow>=10.4.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
huggingface-hub>=0.20.0
pytesseract>=0.3.10
bcrypt>=4.1.0
PyJWT>=2.8.0
